enable use of listmanageActions hook
