+++ .changeset/chilled-eagles-compete.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/app-admin-ui': minor
+---
+
+Enabled use of `listManageActions` hook
