+++ packages/app-admin-ui/client/pages/List/Management.js
@@ -1,15 +1,12 @@
-/* global ENABLE_DEV_FEATURES */
-
-import React, { Fragment, useState } from 'react';
+import React, { Fragment } from 'react';
 import styled from '@emotion/styled';
 
-import { SettingsIcon, TrashcanIcon } from '@arch-ui/icons';
 import { FlexGroup } from '@arch-ui/layout';
-import { IconButton } from '@arch-ui/button';
 import { colors, gridSize } from '@arch-ui/theme';
 
-import UpdateManyItemsModal from '../../components/UpdateManyItemsModal';
-import DeleteManyItemsModal from '../../components/DeleteManyItemsModal';
+import { useUIHooks } from '../../providers/Hooks';
+import DeleteItems from '../../components/DeleteItems';
+import UpdateItems from '../../components/UpdateItems';
 
 export const ManageToolbar = styled.div(({ isVisible }) => ({
   height: 35,
@@ -22,21 +19,8 @@ const SelectedCount = styled.div({
   marginRight: gridSize,
 });
 
-export default function ListManage(props) {
-  const { onDeleteMany, onUpdateMany, selectedItems } = props;
-  const [deleteModalIsVisible, setDeleteModal] = useState(false);
-  const [updateModalIsVisible, setUpdateModal] = useState(false);
-
-  const handleDelete = () => {
-    setDeleteModal(false);
-    onDeleteMany();
-  };
-  const handleUpdate = () => {
-    setUpdateModal(false);
-    onUpdateMany();
-  };
-
-  const { list, pageSize, totalItems } = props;
+export default function ListManage({ pageSize, totalItems, selectedItems }) {
+  const { listManageActions } = useUIHooks();
   const selectedCount = selectedItems.length;
 
   return (
@@ -45,46 +29,15 @@ export default function ListManage(props) {
         <SelectedCount>
           {selectedCount} of {Math.min(pageSize, totalItems)} Selected
         </SelectedCount>
-        {ENABLE_DEV_FEATURES ? (
-          list.access.update ? (
-            <IconButton
-              appearance="primary"
-              icon={SettingsIcon}
-              onClick={() => setUpdateModal(true)}
-              variant="nuance"
-              data-test-name="update"
-            >
-              Update
-            </IconButton>
-          ) : null
-        ) : null}
-        {list.access.update ? (
-          <IconButton
-            appearance="danger"
-            icon={TrashcanIcon}
-            onClick={() => setDeleteModal(true)}
-            variant="nuance"
-            data-test-name="delete"
-          >
-            Delete
-          </IconButton>
-        ) : null}
+        {listManageActions ? (
+          listManageActions()
+        ) : (
+          <Fragment>
+            <UpdateItems />
+            <DeleteItems />
+          </Fragment>
+        )}
       </FlexGroup>
-
-      <UpdateManyItemsModal
-        isOpen={updateModalIsVisible}
-        items={selectedItems}
-        list={list}
-        onClose={() => setUpdateModal(false)}
-        onUpdate={handleUpdate}
-      />
-      <DeleteManyItemsModal
-        isOpen={deleteModalIsVisible}
-        itemIds={selectedItems}
-        list={list}
-        onClose={() => setDeleteModal(false)}
-        onDelete={handleDelete}
-      />
     </Fragment>
   );
 }
