+++ packages/app-admin-ui/client/pages/List/index.js
@@ -26,7 +26,7 @@ import Pagination, { getPaginationLabel } from './Pagination';
 import Search from './Search';
 import Management, { ManageToolbar } from './Management';
 import { useListFilter, useListSelect, useListSort, useListUrlState } from './dataHooks';
-import { useList } from '../../providers/List';
+import { useList, ListDataProvider } from '../../providers/List';
 import { useUIHooks } from '../../providers/Hooks';
 import { useAdminMeta } from '../../providers/AdminMeta';
 import CreateItem from './CreateItem';
@@ -71,16 +71,10 @@ export function ListLayout(props) {
   // Misc.
   // ------------------------------
 
-  const onDeleteSelectedItems = () => {
-    query.refetch();
-    onSelectChange([]);
-  };
   const onDeleteItem = () => {
     query.refetch();
   };
-  const onUpdateSelectedItems = () => {
-    query.refetch();
-  };
+
   const onCreate = ({ data }) => {
     const id = data[list.gqlNames.createMutationName].id;
     query.refetch().then(() => {
@@ -95,155 +89,153 @@ export function ListLayout(props) {
 
   const Render = ({ children }) => children();
   return (
-    <main>
-      <div ref={measureElementRef} />
+    <ListDataProvider {...{ query, items, selectedItems, onSelectChange }}>
+      <main>
+        <div ref={measureElementRef} />
 
-      <Container isFullWidth>
-        <HeaderInset>
-          <FlexGroup align="center" justify="space-between">
-            <PageTitle>{list.plural}</PageTitle>
-            {listHeaderActions ? listHeaderActions() : <CreateItem />}
-          </FlexGroup>
-          <div
-            css={{ alignItems: 'center', display: 'flex', flexWrap: 'wrap' }}
-            id={cypressFiltersId}
-          >
-            <Suspense fallback={<LoadingIndicator css={{ height: '3em' }} size={12} />}>
-              <Render>
-                {() => {
-                  captureSuspensePromises(
-                    fields
-                      .filter(field => field.path !== '_label_')
-                      .map(field => () => field.initCellView())
-                  );
-                  return <Search list={list} isLoading={query.loading} />;
-                }}
-              </Render>
-            </Suspense>
-            <ActiveFilters list={list} />
-          </div>
-
-          <ManageToolbar isVisible css={{ marginLeft: 2 }}>
-            {selectedItems.length ? (
-              <Management
-                list={list}
-                onDeleteMany={onDeleteSelectedItems}
-                onUpdateMany={onUpdateSelectedItems}
-                pageSize={pageSize}
-                selectedItems={selectedItems}
-                onSelectChange={onSelectChange}
-                totalItems={itemCount}
-              />
-            ) : items && items.length ? (
-              <FlexGroup align="center" growIndexes={[0]}>
-                <div
-                  css={{
-                    alignItems: 'center',
-                    color: colors.N40,
-                    display: 'flex',
-                    marginTop: gridSize - 2,
+        <Container isFullWidth>
+          <HeaderInset>
+            <FlexGroup align="center" justify="space-between">
+              <PageTitle>{list.plural}</PageTitle>
+              {listHeaderActions ? listHeaderActions() : <CreateItem />}
+            </FlexGroup>
+            <div
+              css={{ alignItems: 'center', display: 'flex', flexWrap: 'wrap' }}
+              id={cypressFiltersId}
+            >
+              <Suspense fallback={<LoadingIndicator css={{ height: '3em' }} size={12} />}>
+                <Render>
+                  {() => {
+                    captureSuspensePromises(
+                      fields
+                        .filter(field => field.path !== '_label_')
+                        .map(field => () => field.initCellView())
+                    );
+                    return <Search list={list} isLoading={query.loading} />;
                   }}
-                >
-                  <span id="ks-pagination-count">
-                    {getPaginationLabel({
-                      currentPage: currentPage,
-                      pageSize: pageSize,
-                      plural: list.plural,
-                      singular: list.singular,
-                      total: itemCount,
-                    })}
-                    ,
-                  </span>
-                  {sortBy ? (
-                    <Fragment>
-                      <span css={{ paddingLeft: '0.5ex' }}>sorted by</span>
-                      <SortPopout listKey={list.key} />
-                    </Fragment>
-                  ) : (
-                    ''
-                  )}
-                  <span css={{ paddingLeft: '0.5ex' }}>with</span>
-                  <ColumnPopout
-                    listKey={list.key}
-                    target={handlers => (
+                </Render>
+              </Suspense>
+              <ActiveFilters list={list} />
+            </div>
+
+            <ManageToolbar isVisible css={{ marginLeft: 2 }}>
+              {selectedItems.length ? (
+                <Management
+                  pageSize={pageSize}
+                  selectedItems={selectedItems}
+                  totalItems={itemCount}
+                />
+              ) : items && items.length ? (
+                <FlexGroup align="center" growIndexes={[0]}>
+                  <div
+                    css={{
+                      alignItems: 'center',
+                      color: colors.N40,
+                      display: 'flex',
+                      marginTop: gridSize - 2,
+                    }}
+                  >
+                    <span id="ks-pagination-count">
+                      {getPaginationLabel({
+                        currentPage: currentPage,
+                        pageSize: pageSize,
+                        plural: list.plural,
+                        singular: list.singular,
+                        total: itemCount,
+                      })}
+                      ,
+                    </span>
+                    {sortBy ? (
+                      <Fragment>
+                        <span css={{ paddingLeft: '0.5ex' }}>sorted by</span>
+                        <SortPopout listKey={list.key} />
+                      </Fragment>
+                    ) : (
+                      ''
+                    )}
+                    <span css={{ paddingLeft: '0.5ex' }}>with</span>
+                    <ColumnPopout
+                      listKey={list.key}
+                      target={handlers => (
+                        <Button
+                          variant="subtle"
+                          appearance="primary"
+                          spacing="cozy"
+                          id="ks-column-button"
+                          {...handlers}
+                        >
+                          {fields.length} Columns
+                          <DisclosureArrow />
+                        </Button>
+                      )}
+                    />
+                  </div>
+                  <FlexGroup align="center" css={{ marginLeft: '1em' }}>
+                    <Suspense fallback={<LoadingIndicator css={{ height: '3em' }} size={12} />}>
+                      <Render>
+                        {() => {
+                          captureSuspensePromises(
+                            fields
+                              .filter(field => field.path !== '_label_')
+                              .map(field => () => field.initCellView())
+                          );
+                          return <Pagination listKey={list.key} isLoading={query.loading} />;
+                        }}
+                      </Render>
+                    </Suspense>
+                  </FlexGroup>
+                </FlexGroup>
+              ) : null}
+            </ManageToolbar>
+          </HeaderInset>
+        </Container>
+
+        <CreateItemModal onCreate={onCreate} />
+
+        <Container isFullWidth>
+          <ListTable
+            {...props}
+            adminPath={adminPath}
+            columnControl={
+              <ColumnPopout
+                listKey={list.key}
+                target={handlers => (
+                  <Tooltip placement="top" content="Columns">
+                    {ref => (
                       <Button
                         variant="subtle"
-                        appearance="primary"
-                        spacing="cozy"
-                        id="ks-column-button"
+                        css={{
+                          background: 0,
+                          border: 0,
+                          color: colors.N40,
+                        }}
                         {...handlers}
+                        ref={applyRefs(handlers.ref, ref)}
                       >
-                        {fields.length} Columns
-                        <DisclosureArrow />
+                        <KebabHorizontalIcon />
                       </Button>
                     )}
-                  />
-                </div>
-                <FlexGroup align="center" css={{ marginLeft: '1em' }}>
-                  <Suspense fallback={<LoadingIndicator css={{ height: '3em' }} size={12} />}>
-                    <Render>
-                      {() => {
-                        captureSuspensePromises(
-                          fields
-                            .filter(field => field.path !== '_label_')
-                            .map(field => () => field.initCellView())
-                        );
-                        return <Pagination listKey={list.key} isLoading={query.loading} />;
-                      }}
-                    </Render>
-                  </Suspense>
-                </FlexGroup>
-              </FlexGroup>
-            ) : null}
-          </ManageToolbar>
-        </HeaderInset>
-      </Container>
-
-      <CreateItemModal onCreate={onCreate} />
-
-      <Container isFullWidth>
-        <ListTable
-          {...props}
-          adminPath={adminPath}
-          columnControl={
-            <ColumnPopout
-              listKey={list.key}
-              target={handlers => (
-                <Tooltip placement="top" content="Columns">
-                  {ref => (
-                    <Button
-                      variant="subtle"
-                      css={{
-                        background: 0,
-                        border: 0,
-                        color: colors.N40,
-                      }}
-                      {...handlers}
-                      ref={applyRefs(handlers.ref, ref)}
-                    >
-                      <KebabHorizontalIcon />
-                    </Button>
-                  )}
-                </Tooltip>
-              )}
-            />
-          }
-          fields={fields}
-          handleSortChange={handleSortChange}
-          isFullWidth
-          items={items}
-          queryErrors={queryErrors}
-          list={list}
-          onChange={onDeleteItem}
-          onSelectChange={onSelectChange}
-          selectedItems={selectedItems}
-          sortBy={sortBy}
-          currentPage={currentPage}
-          filters={filters}
-          search={search}
-        />
-      </Container>
-    </main>
+                  </Tooltip>
+                )}
+              />
+            }
+            fields={fields}
+            handleSortChange={handleSortChange}
+            isFullWidth
+            items={items}
+            queryErrors={queryErrors}
+            list={list}
+            onChange={onDeleteItem}
+            onSelectChange={onSelectChange}
+            selectedItems={selectedItems}
+            sortBy={sortBy}
+            currentPage={currentPage}
+            filters={filters}
+            search={search}
+          />
+        </Container>
+      </main>
+    </ListDataProvider>
   );
 }
