+++ packages/app-admin-ui/components/index.js
@@ -1,4 +1,4 @@
-export { ListProvider, useList } from '../client/providers/List';
+export { ListProvider, useList, useListData } from '../client/providers/List';
 export { default as ListData, List, ListLayout } from '../client/pages/List/index';
 export { default as ListManage } from '../client/pages/List/Management';
 export { default as FieldSelect } from '../client/pages/List/FieldSelect';
@@ -17,3 +17,5 @@ export { useAdminMeta } from '../client/providers/AdminMeta';
 export { default as AddNewItem } from '../client/pages/Item/AddNewItem';
 export { default as ItemId } from '../client/pages/Item/ItemId';
 export { default as CreateItem } from '../client/pages/List/CreateItem';
+export { default as UpdateItems } from '../client/components/UpdateItems';
+export { default as DeleteItems } from '../client/components/DeleteItems';
